Delete site option 'state_modules' on uninstall
The module manager is creating the site option 'state_modules' on database to manage the modules, but this value is not removed if the plugin is uninstalled. This will cause multiple entries on the database that can generate unexpected behaviours because the get_site_option method return the first entry found on the database.