Extract installation check tasks into separate class
In future we could need more and different things to do on installation
or on version update. Keep all the maintentance in front controller
would quickly mess up it. Moreover, is not rella responsibility of front
controller install or updade the version.

This branch introduces a class to perfrome some checks, and then just
triegger an action. based on the action triggered and on its arguments
things will happen...