Cache the call to get all languages.
Although the result was being memoised and the query is relatively
light, it's still a query which would be called for each request which
can cause problems on large sites with large amounts of traffic.

This change will cache the languages indefinitely until a language is
changed on any site. It uses an md5 of the serialized array of languages
to do this.

This should stop a query being run on each request and addresses part of https://github.com/inpsyde/MultilingualPress/issues/297